refactoring test into a common class
