+++ opentelemetry-api/tests/context/test_threadlocal_context.py
@@ -12,69 +12,22 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import unittest
-from logging import ERROR
 from unittest.mock import patch
 
 from opentelemetry import context
 from opentelemetry.context.threadlocal_context import ThreadLocalRuntimeContext
 
-
-def do_work() -> None:
-    context.attach(context.set_value("say", "bar"))
+from .base_context import ContextTestCases
 
 
-class TestThreadLocalContext(unittest.TestCase):
+class TestThreadLocalContext(ContextTestCases.BaseTest):
     def setUp(self):
-        self.previous_context = context.get_current()
+        super(TestThreadLocalContext, self).setUp()
+        self.mock_runtime = patch.object(
+            context, "_RUNTIME_CONTEXT", ThreadLocalRuntimeContext(),
+        )
+        self.mock_runtime.start()
 
     def tearDown(self):
-        context.attach(self.previous_context)
-
-    @patch(
-        "opentelemetry.context._RUNTIME_CONTEXT",  # type: ignore
-        ThreadLocalRuntimeContext(),
-    )
-    def test_context(self):
-        self.assertIsNone(context.get_value("say"))
-        empty = context.get_current()
-        second = context.set_value("say", "foo")
-
-        self.assertEqual(context.get_value("say", context=second), "foo")
-
-        do_work()
-        self.assertEqual(context.get_value("say"), "bar")
-        third = context.get_current()
-
-        self.assertIsNone(context.get_value("say", context=empty))
-        self.assertEqual(context.get_value("say", context=second), "foo")
-        self.assertEqual(context.get_value("say", context=third), "bar")
-
-    @patch(
-        "opentelemetry.context._RUNTIME_CONTEXT",  # type: ignore
-        ThreadLocalRuntimeContext(),
-    )
-    def test_set_value(self):
-        first = context.set_value("a", "yyy")
-        second = context.set_value("a", "zzz")
-        third = context.set_value("a", "---", first)
-        self.assertEqual("yyy", context.get_value("a", context=first))
-        self.assertEqual("zzz", context.get_value("a", context=second))
-        self.assertEqual("---", context.get_value("a", context=third))
-        self.assertEqual(None, context.get_value("a"))
-
-    @patch(
-        "opentelemetry.context._RUNTIME_CONTEXT",  # type: ignore
-        ThreadLocalRuntimeContext(),
-    )
-    def test_attach(self):
-        context.attach(context.set_value("a", "yyy"))
-
-        token = context.attach(context.set_value("a", "zzz"))
-        self.assertEqual("zzz", context.get_value("a"))
-
-        context.detach(token)
-        self.assertEqual("yyy", context.get_value("a"))
-
-        with self.assertLogs(level=ERROR):
-            context.detach("some garbage")
+        super(TestThreadLocalContext, self).tearDown()
+        self.mock_runtime.stop()
